refactor(gulp): move the js task to a separate file
